old_language_learn_2020/api/pyproject.toml
[tool.poetry]
name = "languagelearn"
version = "0.1.0"
description = ""
authors = ["John Lockwood <johnlockwood2999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^1.3.17"
psycopg2 = "^2.8.5"
pyjwt = "^1.7.1"
flask = "^1.1.2"


[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
